+++ anilist/cache.go
@@ -1,0 +1,3 @@
+package anilist
+
+var cache = make(map[string]*Manga)