+++ provider/custom/translator.go
@@ -5,6 +5,7 @@ import (
 	"github.com/metafates/mangal/source"
 	"github.com/samber/lo"
 	lua "github.com/yuin/gopher-lua"
+	"net/url"
 	"strconv"
 	"strings"
 )
@@ -53,10 +54,22 @@ func mangaFromTable(table *lua.LTable, index uint16) (manga *source.Manga, err e
 	mappings := map[string]mapping{
 		"name":    {A: lua.LTString, B: true, C: func(v string) error { manga.Name = v; return nil }},
 		"url":     {A: lua.LTString, B: true, C: func(v string) error { manga.URL = v; return nil }},
-		"summary": {A: lua.LTString, B: false, C: func(v string) error { manga.Summary = v; return nil }},
-		"author":  {A: lua.LTString, B: false, C: func(v string) error { manga.Author = v; return nil }},
+		"summary": {A: lua.LTString, B: false, C: func(v string) error { manga.Metadata.Summary = v; return nil }},
+		"cover": {A: lua.LTString, B: false, C: func(v string) error {
+			if v == "" {
+				return nil
+			}
+
+			_, err := url.Parse(v)
+			if err != nil {
+				return err
+			}
+
+			manga.Metadata.Cover = v
+			return nil
+		}},
 		"genres": {A: lua.LTString, B: false, C: func(v string) error {
-			manga.Genres = lo.Map(strings.Split(v, ","), func(genre string, _ int) string {
+			manga.Metadata.Genres = lo.Map(strings.Split(v, ","), func(genre string, _ int) string {
 				return strings.TrimSpace(genre)
 			})
 			return nil
@@ -78,12 +91,23 @@ func chapterFromTable(table *lua.LTable, manga *source.Manga, index uint16) (cha
 		"name":          {A: lua.LTString, B: true, C: func(v string) error { chapter.Name = v; return nil }},
 		"url":           {A: lua.LTString, B: true, C: func(v string) error { chapter.URL = v; return nil }},
 		"volume":        {A: lua.LTString, B: false, C: func(v string) error { chapter.Volume = v; return nil }},
-		"manga_summary": {A: lua.LTString, B: false, C: func(v string) error { manga.Summary = v; return nil }},
-		"manga_author":  {A: lua.LTString, B: false, C: func(v string) error { manga.Author = v; return nil }},
+		"manga_summary": {A: lua.LTString, B: false, C: func(v string) error { manga.Metadata.Summary = v; return nil }},
 		"manga_genres": {A: lua.LTString, B: false, C: func(v string) error {
-			manga.Genres = lo.Map(strings.Split(v, ","), func(genre string, _ int) string {
+			manga.Metadata.Genres = lo.Map(strings.Split(v, ","), func(genre string, _ int) string {
 				return strings.TrimSpace(genre)
 			})
+			return nil
+		}},
+		"manga_cover": {A: lua.LTString, B: false, C: func(v string) error {
+			if v == "" {
+				return nil
+			}
+			_, err := url.Parse(v)
+			if err != nil {
+				return err
+			}
+
+			manga.Metadata.Cover = v
 			return nil
 		}},
 	}